// packages/client/src/utils/knownStrings.ts
export enum KnownStrings {
    ComponentWorkflowId = "wfId",
    XtensionInput = "['X_NTX_XTENSION_INPUT']",
    NTXConnectionId = "NTX_CONNECTION_ID",
    ContractIdToken = "<<CONTRACT_ID>>",
    ConnectionIdToken = "<<CONNECTION_ID>>",
    ConnectionToken = "<<CONNECTION>>",
    ContractToken = "<<CONTRACT>>"
}